FinanceApp/stockProject - main/stockProject/requirements.txt
Django
djangorestframework
gunicorn
celery
redis
django-celery-beat
gevent
kombu
drf-yasg
psycopg2-binary